+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0
+
+* __BREAKING__ rename create => creator
+* __BREAKING__ remove StateScope.value()
+* added test
+
 ## 1.0.2
 
 * Changed description.